requirements.txt
numpy==1.16.0
Pillow==6.1.0
scipy==1.3.1
tensorflow-gpu==0.12.1
easydict==1.9
matplotlib==3.1.1
Cython==0.29.13
Image
opencv-python==4.1.1.26
pyyaml==5.1.2
scikit-image==0.15.0
scikit-learn==0.21.3
seaborn==0.9.0